+++ app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 
 import { Inter } from "next/font/google";
 import "./globals.css";
-import { Providers } from "@/app/providers";
+import { Providers } from "@/providers";
 
 const inter = Inter({ subsets: ["latin"] });
