+++ providers/config.ts
@@ -2,12 +2,12 @@ import { createConfig, fallback, http } from "wagmi";
 import { getDefaultConfig } from "connectkit";
 import { gnosis, mainnet } from "wagmi/chains";
 import { safe } from "wagmi/connectors";
-import { ChainId, RPC_LIST } from "@/app/constants";
+import { ChainId, RPC_LIST } from "@/constants";
 
 const defaultConfig = getDefaultConfig({
   chains: [mainnet, gnosis],
   walletConnectProjectId:
-    process.env.NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID || "",
+    process.env.NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID ?? "",
   transports: {
     [mainnet.id]: fallback([http(RPC_LIST[ChainId.ETHEREUM]), http()]),
     [gnosis.id]: fallback([http(RPC_LIST[ChainId.GNOSIS]), http()]),
